Changement des Double Id en int id

+++ src/main/java/org/example/proxishop/controller/ProductController.java
@@ -71,7 +71,7 @@ public class ProductController {
      * @throws SQLException Si une erreur SQL se produit.
      */
     @PostMapping("/addProducts")
-    public String addProducts(@RequestParam double subCategoryid, @RequestParam String bddname, @RequestParam String productName,
+    public String addProducts(@RequestParam int subCategoryid, @RequestParam String bddname, @RequestParam String productName,
                               @RequestParam String description, @RequestParam double price, @RequestParam double stock, @RequestParam String image,
                               Model model, RedirectAttributes redirectAttributes) throws SQLException {
         BdProducts db = new BdProducts();
@@ -97,7 +97,7 @@ public class ProductController {
      * @throws SQLException Si une erreur SQL se produit.
      */
     @PostMapping("/updateProducts")
-    public String updateProducts(@RequestParam double subCategoryid, @RequestParam String bddname, @RequestParam String productName,
+    public String updateProducts(@RequestParam int subCategoryid, @RequestParam String bddname, @RequestParam String productName,
                                  @RequestParam String description, @RequestParam double price, @RequestParam double stock, @RequestParam String image,
                                  @RequestParam int id_product, Model model, RedirectAttributes redirectAttributes) throws SQLException {
         BdProducts db = new BdProducts();
