+++ src/main/java/org/example/proxishop/model/entities/proxi/Shopkeepers.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 public class Shopkeepers {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Double id;
+    private int id;
     private Double siret;
     private String firstName;
     private String lastName;
@@ -19,13 +19,13 @@ public class Shopkeepers {
     private String webSiteName;
     private String firmName;
     private String adress;
-    private Double id_offer;
+    private int id_offer;
 
 
     public Shopkeepers() {
     }
 
-    public Shopkeepers(Double siret, String firstName, String lastName, String email, String webSiteName, String firmName, String adress, Double id_offer) {
+    public Shopkeepers(Double siret, String firstName, String lastName, String email, String webSiteName, String firmName, String adress, int id_offer) {
         this.siret = siret;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -36,7 +36,7 @@ public class Shopkeepers {
         this.id_offer = id_offer;
     }
 
-    public Shopkeepers(Double id, Double siret, String firstName, String lastName, String email, String webSiteName, String firmName, String adress, Double id_offer) {
+    public Shopkeepers(int id, Double siret, String firstName, String lastName, String email, String webSiteName, String firmName, String adress, int id_offer) {
         this.id = id;
         this.siret = siret;
         this.firstName = firstName;
